==> vignettes/Tgondii_screen/01.2_plot_plate_maps.Rmd <==
---
title: "02.2_plot_plate_maps"
output: html_document
---

```{r load-data}

load("intermediate_data/data_gene.Rdata")
```


```{r positivity-by-plate-layout}

plot_data <- data_gene |>
  dplyr::group_by(plate, row, column) |>
  dplyr::summarize(
    maxvit_positive = maxvit_positive[1],
    maxvit_negative = maxvit_negative[1],
    positivity_rate =
      maxvit_positive[1] / (maxvit_positive[1] + maxvit_negative[1]),
      .groups = "drop") |>
  dplyr::mutate(
    plate_label = paste0("Plate ", plate),
    row_label = paste0("R", row),
    column_label = paste0("C", column))

ggplot2::ggplot() +
  ggplot2::theme_bw() +
  ggplot2::geom_tile(
    data = plot_data,
    mapping = ggplot2::aes(
      x = column,
      y = row,
      fill = positivity_rate)) +
  ggplot2::facet_wrap(facets = dplyr::vars(plate_label)) +
  ggplot2::scale_x_continuous(
    "Column",
    labels = 1:24, breaks = 1:24,
    expand = c(0, 0)) +
  ggplot2::scale_y_reverse(
    "Row",
    labels = 1:16, breaks = 1:16,
    expand = c(0, 0)) +
  ggplot2::scale_fill_continuous("Positivity Rate") +
  ggplot2::theme(legend.position = "bottom")


ggplot2::ggsave(
  filename = "product/plate_map_positivity_rate_20250618.pdf",
  width = 7.5,
  height = 5)


```


```{r n-genes-by-plate-layout}

gene_in_well_threshold <- 2/100
plot_data <- data_gene |>
  dplyr::filter(well_type == "Screen") |>
  dplyr::mutate(
    gene_in_well = well_gene_fraction >= gene_in_well_threshold) |>
  dplyr::group_by(plate, row, column) |>
  dplyr::summarize(
    n_genes = sum(gene_in_well > 0),
    .groups = "drop") |>
  dplyr::bind_rows(
    tidyr::expand_grid(
      plate = 1:4, row = 1:16, column = 1:3,
      n_genes = NA)) |>
  dplyr::mutate(
    plate_label = paste0("Plate ", plate))
  

ggplot2::ggplot() +
  ggplot2::theme_bw() +
  ggplot2::geom_tile(
    data = plot_data,
    mapping = ggplot2::aes(
      x = column,
      y = row,
      fill = sqrt(n_genes))) +
  ggplot2::facet_wrap(facets = dplyr::vars(plate_label)) +
  ggplot2::scale_x_continuous(
    "Column",
    labels = 1:24, breaks = 1:24,
    expand = c(0, 0)) +
  ggplot2::scale_y_reverse(
    "Row",
    labels = 1:16, breaks = 1:16,
    expand = c(0, 0)) +
  ggplot2::scale_fill_viridis_c(
    "N Genes Per Well",
    limits = sqrt(c(0, 20)),
    labels = c(0, 1, 4, 9, 16),
    breaks = sqrt(c(0, 1, 4, 9, 16)),
    option = "C") +
  ggplot2::theme(legend.position = "bottom")


ggplot2::ggsave(
  filename = "product/plate_map_n_genes_20250618.pdf",
  width = 7.5,
  height = 5)


```
